'use strict';
Added to resumeBuilder.js and worked fine, the validator warned to use
it in function form.

Tried with helper.js as well, but it messed up the program working
properly, and the validator showed no errors, only the one
aforementioned warning.  Tried function form as well, with problematic
results.

More learning is clearly needed here.