Temp Map Locations Solution
Google Maps only allows 10 locations for markers even if they are
duplicates.  As a result the other locations do not get  recognized.  I
commented out some of the ForEach loops to reduce the number of times
Salt Lake City appears.  This allows the remaining locations to be
marked on the map.  A better solution will be to remove duplicated
location strings from the locations array before google place services
uses them.